Added SSR to home and planet details page, refactored these two pages/components
